# openapi.yaml
openapi: 3.0.3
info:
  title: API Tablero Cooptech
  version: 1.0.0
  description: API REST del Tablero de Mando Cooptech. Single-tenant, uso interno.
    Autenticación delegada al servicio de identidad central (Bearer JWT).
servers:
- url: /api/v1
security:
- bearerAuth: []
...
paths:
...
  /guardias:
    get:
      tags:
      - guardias
      summary: Listar guardias
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Guardia'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - guardias
      summary: Crear guardia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardiaInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guardia'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /guardias/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      tags:
      - guardias
      summary: Obtener por id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guardia'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - guardias
      summary: Actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardiaInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guardia'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Validation'
    delete:
      tags:
      - guardias
      summary: Eliminar
      responses:
        '204':
          description: Eliminado
        '404':
          $ref: '#/components/responses/NotFound'
...
  /archivos:
    post:
      tags:
      - archivos
      summary: Subir archivo a MinIO
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                leadId: &id002
                  type: integer
                contexto: &id004
                  type: string
                gpsLat: &id001
                  type: number
                gpsLng: *id001
                esBoceto: &id005
                  type: boolean
      responses:
        '201':
          description: Subido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Archivo'
  /archivos/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      tags:
      - archivos
      summary: Metadatos del archivo
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Archivo'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - archivos
      summary: Eliminar archivo (de la base y de MinIO)
      responses:
        '204':
          description: Eliminado
        '404':
          $ref: '#/components/responses/NotFound'
  /grilla:
    get:
      tags:
      - grilla
      summary: Entradas de grilla por rango/colaborador
      parameters:
      - name: colaboradorId
        in: query
        schema: *id002
      - name: desde
        in: query
        schema: &id003
          type: string
          format: date
      - name: hasta
        in: query
        schema: *id003
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrillaEntrada'
    put:
      tags:
      - grilla
      summary: Crear o actualizar la entrada de un día (upsert por colaborador+fecha)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrillaEntradaInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrillaEntrada'
...
  /costos/{mes}:
    parameters:
    - name: mes
      in: path
      required: true
      schema:
        type: string
        example: 2026-06
    get:
      tags:
      - costos
      summary: Costos de un mes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostoMensual'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - costos
      summary: Cargar/actualizar costos del mes (rol gerencial)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostoMensualInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostoMensual'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: *id002
...
  responses:
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Validation:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *id004
            message: *id004
...
    Guardia:
      type: object
      properties:
        id: *id002
        colaboradorId: *id002
        fechaInicio: *id003
        fechaFin: &id049
          type: string
          format: date
          nullable: true
        notas: &id050
          type: string
          nullable: true
    GuardiaInput:
      type: object
      properties:
        colaboradorId: *id002
        fechaInicio: *id003
        fechaFin: *id049
        notas: *id050
...
    Archivo:
      type: object
      properties:
        id: *id002
        key: *id004
        url:
          type: string
          nullable: true
        nombre: *id004
        mime:
          type: string
          nullable: true
        tamano:
          type: integer
          nullable: true
        gpsLat:
          type: number
          nullable: true
        gpsLng:
          type: number
          nullable: true
        esBoceto: *id005
        leadId:
          type: integer
          nullable: true
        contexto:
          type: string
          nullable: true
    GrillaEntrada:
      type: object
      properties:
        id: *id002
        colaboradorId: *id002
        fecha: *id003
        estado:
          type: string
          enum:
          - present
          - home_office
          - vacaciones
          - franco
          - franco_cumple
          - feriado
          - licencia
          - viaje
        items:
          type: array
          items:
            type: object
          nullable: true
    GrillaEntradaInput:
      type: object
      properties:
        colaboradorId: *id002
        fecha: *id003
        estado:
          type: string
          enum:
          - present
          - home_office
          - vacaciones
          - franco
          - franco_cumple
          - feriado
          - licencia
          - viaje
        items:
          type: array
          items:
            type: object
          nullable: true
      required:
      - colaboradorId
      - fecha
    CostoMensual:
      type: object
      properties:
        id: *id002
        mes: *id004
        costoLaboral:
          type: number
          nullable: true
        cotizacionDolar:
          type: number
          nullable: true
    CostoMensualInput:
      type: object
      properties:
        costoLaboral:
          type: number
          nullable: true
        cotizacionDolar:
          type: number
          nullable: true