# src/routes/archivos.js
import { Router } from 'express';
import { prisma } from '../lib/prisma.js';
import { ApiError } from '../middleware/errorHandler.js';

const router = Router();

// Almacenamiento de archivos:
// El binario lo sube y lo lee directamente el FRONTEND contra el gateway de
// almacenamiento (storageov → MinIO). Acá sólo guardamos la REFERENCIA: el
// `key` (nombre del archivo en el bucket) junto con su metadata y el contexto
// (objetivo / lead) al que pertenece. La imagen se resuelve en el cliente con
// ese `key` mediante getImage().

// Listar referencias, filtrando por objetivo o lead (?objetivoId= / ?leadId=).
router.get('/', async (req, res, next) => {
  try {
    const where = {};
    if (req.query.objetivoId) where.objetivoId = Number(req.query.objetivoId);
    if (req.query.leadId) where.leadId = Number(req.query.leadId);
    if (req.query.contexto) where.contexto = String(req.query.contexto);
    const data = await prisma.archivo.findMany({ where, orderBy: { createdAt: 'asc' } });
    res.json({ data });
  } catch (e) { next(e); }
});

// Registrar la referencia de un archivo que el frontend ya subió al gateway.
// Body (JSON): { key, nombre?, mime?, tamano?, gpsLat?, gpsLng?, esBoceto?,
//                leadId?, objetivoId?, contexto?, url? }
router.post('/', async (req, res, next) => {
  try {
    const b = req.body || {};
    if (!b.key) throw new ApiError(422, 'validation', 'Falta "key" (nombre del archivo en el almacenamiento)');
    const archivo = await prisma.archivo.create({
      data: {
        key: String(b.key),
        url: b.url || null,
        nombre: b.nombre || String(b.key),
        mime: b.mime || null,
        tamano: b.tamano != null && b.tamano !== '' ? Number(b.tamano) : null,
        gpsLat: b.gpsLat != null && b.gpsLat !== '' ? Number(b.gpsLat) : null,
        gpsLng: b.gpsLng != null && b.gpsLng !== '' ? Number(b.gpsLng) : null,
        esBoceto: b.esBoceto === true || b.esBoceto === 'true',
        leadId: b.leadId != null && b.leadId !== '' ? Number(b.leadId) : null,
        objetivoId: b.objetivoId != null && b.objetivoId !== '' ? Number(b.objetivoId) : null,
        contexto: b.contexto || null,
      },
    });
    res.status(201).json(archivo);
  } catch (e) { next(e); }
});

router.get('/:id', async (req, res, next) => {
  try {
    const a = await prisma.archivo.findUnique({ where: { id: Number(req.params.id) } });
    if (!a) throw new ApiError(404, 'not_found', 'Archivo no encontrado');
    res.json(a);
  } catch (e) { next(e); }
});

router.delete('/:id', async (req, res, next) => {
  try {
    const a = await prisma.archivo.findUnique({ where: { id: Number(req.params.id) } });
    if (!a) throw new ApiError(404, 'not_found', 'Archivo no encontrado');
    // Sólo borramos la referencia en la base. El gateway de almacenamiento no
    // expone (todavía) un endpoint de borrado; si más adelante lo agrega, el
    // borrado del binario lo hace el frontend o se implementa aquí.
    await prisma.archivo.delete({ where: { id: a.id } });
    res.status(204).end();
  } catch (e) { next(e); }
});

export default router;